+++ src/main/java/lotto/service/LottoService.java
@@ -1,28 +1,16 @@
 package lotto.service;
 
 import lotto.domain.Lotto;
-import lotto.repository.LottoRepository;
-import lotto.util.LottoGenerator;
+import lotto.domain.LottoResultDto;
+import lotto.domain.WinningLotto;
 
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
-public class LottoService {
-    private final LottoRepository lottoRepository;
+public interface LottoService {
+    public void buyLottos(Integer money);
 
-    public LottoService(LottoRepository lottoRepository) {
-        this.lottoRepository = lottoRepository;
-    }
+    public List<Lotto> findAll();
 
-    public void buyLottos(Long money) {
-        long count = money / 1000;
-        while (count != 0) {
-            List<Integer> numbers = LottoGenerator.generateLotto();
-            Collections.sort(numbers);
-            Lotto lotto = new Lotto(numbers);
-            lottoRepository.save(lotto);
-            count--;
-        }
-    }
+    public LottoResultDto calculateLottos(WinningLotto winningLotto);
+
 }
